Avoid a Divide by Zero in Momentum Solvers
In the momentum equation solvers, a divide by zero takes place on any dry node (or potentially if "DDU" = 0 on a wet node, although this technically should not happen). Since for dry nodes in particular, the expressions for UU2 and VV2 (or QX2 and QY2) become zero divided by zero, the velocity solution becomes NaNs, which leads to difficulties in obtaining the lateral stresses on the timestep when an element is wetted, since these depend on UU1 and VV1, which will be NaNs for any previously dry node.

We now check for wet/dry state or for DDU = 0 before performing the final solve for the momentum equations, to avoid this situation. If the node is dry, or if DDU = 0, then the solution is also set to 0.